refactor: 💡 allow target connect without read ability
