+++ ui/desktop/app/templates/scopes/scope/projects/targets/index.hbs
@@ -97,7 +97,6 @@
     {{/if}}
 
     {{#if @model}}
-
       <Hds::Table
         @model={{@model}}
         @columns={{array
@@ -160,22 +159,29 @@
             </B.Td>
             <B.Td>
               {{#if (can 'connect target' B.data)}}
-                <Hds::Button
-                  data-test-targets-connect-button={{B.data.id}}
-                  @text={{t 'resources.session.actions.connect'}}
-                  @color='secondary'
-                  @route='scopes.scope.projects.targets.target'
-                  @model={{B.data.id}}
-                  @query={{hash isConnecting=true}}
-                />
+                {{#if (can 'read target' B.data)}}
+                  <Hds::Button
+                    data-test-targets-connect-button={{B.data.id}}
+                    @text={{t 'resources.session.actions.connect'}}
+                    @color='secondary'
+                    @route='scopes.scope.projects.targets.target'
+                    @model={{B.data.id}}
+                    @query={{hash isConnecting=true}}
+                  />
+                {{else}}
+                  <Hds::Button
+                    data-test-targets-connect-button={{B.data.id}}
+                    @text={{t 'resources.session.actions.connect'}}
+                    @color='secondary'
+                    {{on 'click' (fn this.quickConnect B.data)}}
+                  />
+                {{/if}}
               {{/if}}
             </B.Td>
           </B.Tr>
         </:body>
       </Hds::Table>
-
     {{else}}
-
       <Rose::Layout::Centered>
         <Rose::Message
           @title={{t 'resources.target.messages.none.title'}}
@@ -186,7 +192,6 @@
           </message.description>
         </Rose::Message>
       </Rose::Layout::Centered>
-
     {{/if}}
   </page.body>
